requirements.txt
requests==2.32.3
pyyaml==6.0.2
load_dotenv==0.1.0
